refactor: move duplicate detector to utils
