+++ packages/trackerless-network/src/logic/RandomGraphNode.ts
@@ -14,7 +14,7 @@ import { RemoteRandomGraphNode } from './RemoteRandomGraphNode'
 import { INetworkRpc } from '../proto/packages/trackerless-network/protos/NetworkRpc.server'
 import { Empty } from '../proto/google/protobuf/empty'
 import { ServerCallContext } from '@protobuf-ts/runtime-rpc'
-import { DuplicateMessageDetector, NumberPair } from './DuplicateMessageDetector'
+import { DuplicateMessageDetector, NumberPair } from '@streamr/utils'
 
 export enum Event {
     MESSAGE = 'streamr:layer2:random-graph-node:onmessage'
