+++ packages/utils/test/NumberPair.test.ts
@@ -1,4 +1,4 @@
-import { NumberPair } from '../../src/logic/DuplicateMessageDetector'
+import { NumberPair } from '../src/DuplicateMessageDetector'
 
 test('equalTo', () => {
     expect(new NumberPair(5, 2).equalTo(new NumberPair(5, 3))).toEqual(false)
